Move smallrye-config-source-zookeeper to io.smallrye.config.source.zookeeper
