+++ sources/zookeeper/src/main/resources/META-INF/services/org.eclipse.microprofile.config.spi.ConfigSource
@@ -1,1 +1,1 @@
-io.smallrye.configsource.ZooKeeperConfigSource
+io.smallrye.config.source.zookeeper.ZooKeeperConfigSource
